Data is parsed and contained
Data is all parsed properly and it is fully functional. We also saved all the data in a player object. We saved all the player objects in a list of player objects. Now all that is left is to properly get that average, and make a way to specifically get a players information.